# ai-service/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Enhanced ReAct Agent API
  description: |
    Production-ready AI Agent API with advanced features including tool execution, 
    conversation memory, authentication, and real-time streaming capabilities.
    
    ## Features
    - 🤖 AI-powered query processing with ReAct architecture
    - 🔧 Extensible tool system for real-time data access
    - 💬 Conversation memory and context management
    - 🔐 JWT & API Key authentication
    - 📊 Export functionality (JSON, Markdown, PDF)
    - ⚡ WebSocket streaming support
    - 📈 Performance metrics and monitoring
    
    ## Authentication
    The API supports two authentication methods:
    1. **JWT Token**: For user sessions with full access
    2. **API Key**: For programmatic access with rate limiting
  version: 2.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
...
paths:
...
  /cache/stats:
    get:
      tags:
        - Admin
      summary: Get cache statistics
      description: Retrieves cache performance statistics
      operationId: getCacheStats
      security:
        - bearerAuth: []
        - apiKey: []
      responses:
        '200':
          description: Cache statistics retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details

    # Authentication Schemas
...
    CacheStatsResponse:
      type: object
      properties:
        success:
          type: boolean
        stats:
          type: object
          properties:
            hits:
              type: integer
            misses:
              type: integer
            hitRate:
              type: number
            size:
              type: integer
            maxSize:
              type: integer
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Authentication required
...